requirements.txt
requests>=2.25.1
beautifulsoup4>=4.9.3
lxml>=4.6.3
pillow>=8.2.0
python-dotenv>=0.19.0